+++ include/ECS/components/Inertia_comp.hpp
@@ -4,6 +4,6 @@ struct Inertia_Component {
     
     int entity_id;
 
-    float I;
+    double I;
 
 };
