+++ include/ECS/components/Torque_comp.hpp
@@ -4,6 +4,6 @@ struct Torque_Component {
 
     int entity_id;
 
-    float torque;  
+    double torque;  
 
 };
