+++ src/Controller.cpp
@@ -14,9 +14,9 @@ void Controller_System(ECS_Manager &world){
     for (auto it = world.get_component_begin<Controller_Component>();
               it < world.get_component_end<Controller_Component>(); it++){
         
-        Eigen::Vector2f des_pos = it->input;
+        Eigen::Vector2d des_pos = it->input;
 
-        Eigen::Vector2f act_pos = world.get_component<Position_Component>(it->entity_id)->position;
+        Eigen::Vector2d act_pos = world.get_component<Position_Component>(it->entity_id)->position;
         world.get_component<Acceleration_Component>(it->entity_id)->accel = it->K_Gain*(des_pos - act_pos);
     
     }
